beginning to buid up TreeFillRateLimiter

+++ src/main/java/com/duprasville/limiters/RateLimiter.java
@@ -1,10 +1,10 @@
 package com.duprasville.limiters;
 
 public interface RateLimiter {
-    boolean tryAcquire(int permits);
+    boolean tryAcquire(long permits);
 
     default boolean tryAcquire() {
-        return tryAcquire(1);
+        return tryAcquire(1L);
     }
 
     default void setRate(long permitsPerSecond) {
